fixes for cis testts

+++ src/snapred/backend/recipe/algorithm/WashDishes.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Tuple
 import numpy as np
 from mantid.api import AlgorithmFactory, PythonAlgorithm
 from mantid.kernel import Direction, StringArrayProperty
-from mantid.simpleapi import DeleteWorkspace, DeleteWorkspaces
+from mantid.simpleapi import DeleteWorkspace, DeleteWorkspaces, mtd
 
 from snapred.meta.Config import Config
 
@@ -28,11 +28,11 @@ class WashDishes(PythonAlgorithm):
     def PyExec(self) -> None:
         self.log().notice("Washing the dishes...")
         workspace = self.getProperty("Workspace").value
-        if workspace != "" and not self._CISmode:
+        if workspace != "" and mtd.doesExist(workspace) and not self._CISmode:
             DeleteWorkspace(Workspace=workspace)
         workspaces = self.getProperty("WorkspaceList").value
         if workspaces != [] and not self._CISmode:
-            workspaces = [w for w in workspaces if isinstance(w, str)]
+            workspaces = [w for w in workspaces if isinstance(w, str) and mtd.doesExist(w)]
             DeleteWorkspaces(workspaces)
